+++ examples/ipex/text-generation/run_generation.py
@@ -1,7 +1,7 @@
 import torch
 from transformers import AutoTokenizer, pipeline
 
-from optimum.intel.ipex.modeling_decoder import IPEXModelForCausalLM, IPEXModelForSequenceClassification
+from optimum.intel.ipex.modeling_decoder import IPEXModelForCausalLM
 
 
 model_id = "gpt2"
@@ -10,8 +10,3 @@ tokenizer = AutoTokenizer.from_pretrained(model_id)
 text_generator = pipeline("text-generation", model=model, tokenizer=tokenizer)
 
 print(text_generator("This is an example input"))
-
-
-model = IPEXModelForSequenceClassification.from_pretrained(model_id, export=True, torch_dtype=torch.bfloat16)
-text_classifer = pipeline("text-classification", model=model, tokenizer=tokenizer)
-print(text_classifer("This movie is disgustingly good !"))
