+++ optimum/intel/ipex/__init__.py
@@ -1,1 +1,4 @@
 from .inference import inference_mode
+
+
+generation_tasks = "text-generation"
