+++ optimum/intel/ipex/modeling_base.py
@@ -4,31 +4,34 @@ from pathlib import Path
 from tempfile import TemporaryDirectory
 from typing import Optional, Union
 
+import intel_extension_for_pytorch as ipex
 import torch
 from huggingface_hub import hf_hub_download
 from transformers import (
+    AutoConfig,
     AutoModel,
+    AutoModelForSequenceClassification,
     GenerationMixin,
     PretrainedConfig,
-    AutoConfig,
 )
 from transformers.models.auto.auto_factory import _get_model_class
 from transformers.utils import WEIGHTS_NAME
 
-from optimum.modeling_base import OptimizedModel
-from ..utils.import_utils import is_torch_version
 from optimum.exporters import TasksManager
-import intel_extension_for_pytorch as ipex
+from optimum.modeling_base import OptimizedModel
+
+from ..utils.import_utils import is_torch_version
 from ..utils.modeling_utils import patch_decoder_attention_mask
+from . import generation_tasks
+
 
 SUPPORT_MODEL_LIST_FOR_CAUSAL_LM = {
-  #  "llama": LlamaForCausalLM
+    #  "llama": LlamaForCausalLM
 }
 
-SUPPORT_TASK_LIST = {
-    "text-generation": SUPPORT_MODEL_LIST_FOR_CAUSAL_LM
-}
+SUPPORT_TASK_LIST = {"text-generation": SUPPORT_MODEL_LIST_FOR_CAUSAL_LM}
 from ..generation.modeling import jit_trace
+
 
 logger = logging.getLogger(__name__)
 
@@ -46,14 +49,12 @@ class IPEXModel(OptimizedModel):
         use_cache: bool = True,
         **kwargs,
     ):
-        OptimizedModel.__init__(self,
-            model=model, config=config
-        )
-        # add XPU support
+        OptimizedModel.__init__(self, model=model, config=config)
+        # To do: add XPU support
         self._device = torch.device("cpu")
         self.model.to(self._device)
 
-        # Registers the INCModelForXXX classes into the transformers AutoModel classes to avoid warnings when creating
+        # Registers the IPEXModelForXXX classes into the transformers AutoModel classes to avoid warnings when creating
         # a pipeline https://github.com/huggingface/transformers/blob/cad61b68396a1a387287a8e2e2fef78a25b79383/src/transformers/pipelines/base.py#L863
         AutoConfig.register(self.base_model_prefix, AutoConfig)
         if hasattr(self.auto_model_class, "register"):
@@ -88,6 +89,8 @@ class IPEXModel(OptimizedModel):
             "torch_dtype": torch_dtype,
             "device": "cpu",
         }
+        if task not in generation_tasks:
+            model_kwargs.pop("use_cache")
         model_type = None
         support_ipex_transformers = False
         if task in SUPPORT_TASK_LIST.keys():
@@ -106,7 +109,7 @@ class IPEXModel(OptimizedModel):
 
         model = ipex.optimize(model, dtype=torch_dtype, level="O1", auto_kernel_selection=True)
 
-        if kwargs.pop("jit",True):
+        if kwargs.pop("jit", True):
             try:
                 traced_model = cls.apply_jit_optimize(model, task, use_cache, support_ipex_transformers)
                 save_dir = TemporaryDirectory()
@@ -218,3 +221,8 @@ class IPEXModel(OptimizedModel):
     @classmethod
     def apply_jit_optimize(cls, model, task, use_cache, support_ipex_transformers=False):
         return jit_trace(model, task, use_cache)
+
+
+class IPEXModelForSequenceClassification(IPEXModel):
+    auto_model_class = AutoModelForSequenceClassification
+    export_feature = "text-classification"
